refactor escalas logic
